requirements.txt
# Core Data Science Libraries
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.3.0

# Machine Learning Libraries
xgboost>=1.7.0
lightgbm>=3.3.0

# Deep Learning Libraries
tensorflow>=2.12.0
torch>=2.0.0
torchvision>=0.15.0

# Time Series Libraries
prophet>=1.1.0
statsmodels>=0.14.0

# Database and API
supabase>=1.0.0
python-dotenv>=1.0.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
joblib>=1.3.0
tqdm>=4.65.0
